// src/payload-types.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file was automatically generated by Payload.
 * Do not modify it by hand. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export interface User {
  id: string;
  email: string;
  role?: 'admin' | 'editor' | 'reporter' | 'reader' | null;
  name?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Media {
  id: string;
  alt: string;
  caption?: any | null;
  credit: string;
  source: string;
  licenseType: 'public_domain' | 'creative_commons' | 'licensed' | 'fair_use' | 'original';
  expiryDate?: string | null;
  transcript?: string | null;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Story {
  id: string;
  headline: string;
  slug: string;
  strap?: string | null;
  caption?: string | null;
  heroMedia?: string | Media | null;
  layout: any;
  section?: 'accountability' | 'investigative' | 'feature' | 'news' | 'opinion' | any | null;
  status?: 'draft' | 'pending_review' | 'published' | null;
  createdAt: string;
  updatedAt: string;
}